Also remove some gotos and move code into an anonymous namespace.

And then do the partial revert to avoid size_t going negative.

R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1356373003 .
Review URL: https://codereview.chromium.org/1400723002 .

(cherry picked from commit d03ba8d1a5928e8f3f6bd7da063b53b0bc40abfd)
(cherry picked from commit a398ca611d1925182ff6a4e8b2b43c277c364dc0)

Review URL: https://codereview.chromium.org/1406373002 .

+++ core/src/fpdfapi/fpdf_page/fpdf_page_parser.cpp
@@ -71,7 +71,7 @@ CPDF_StreamContentParser::CPDF_StreamContentParser(
     m_pCurStates->m_TextState.New();
     m_pCurStates->m_ColorState.New();
   }
-  for (int i = 0; i < FX_ArraySize(m_Type3Data); ++i) {
+  for (size_t i = 0; i < FX_ArraySize(m_Type3Data); ++i) {
     m_Type3Data[i] = 0.0;
   }
 }
